requirements.txt
django
djangorestframework
coreapi